gcp/requirements.txt
google-cloud-storage
nltk
scikit-learn